+++ pgdog/src/frontend/client/mod.rs
@@ -164,24 +164,25 @@ impl Client {
         // the auth exchange.  MD5, SCRAM, and plain all need the actual
         // password, so this must happen before the first message is sent.
         let resolved;
-        let passwords: &[PasswordKind] =
-            if passwords.iter().any(|p| matches!(p, PasswordKind::VaultStaticRole(_))) {
-                let mut buf = Vec::with_capacity(passwords.len());
-                for p in passwords {
-                    match p {
-                        PasswordKind::VaultStaticRole(path) => {
-                            let pw =
-                                crate::backend::auth::vault::static_client_password(path).await?;
-                            buf.push(PasswordKind::Plain(pw));
-                        }
-                        other => buf.push(other.clone()),
+        let passwords: &[PasswordKind] = if passwords
+            .iter()
+            .any(|p| matches!(p, PasswordKind::VaultStaticRole(_)))
+        {
+            let mut buf = Vec::with_capacity(passwords.len());
+            for p in passwords {
+                match p {
+                    PasswordKind::VaultStaticRole(path) => {
+                        let pw = crate::auth::vault::static_client_password(path).await?;
+                        buf.push(PasswordKind::Plain(pw));
                     }
+                    other => buf.push(other.clone()),
                 }
-                resolved = buf;
-                &resolved
-            } else {
-                passwords
-            };
+            }
+            resolved = buf;
+            &resolved
+        } else {
+            passwords
+        };
 
         let result = match auth_type {
             AuthType::Md5 => {
